refactor: 設定にモデルモジュールを追加
