+++ run.py
@@ -16,11 +16,8 @@ from voicevox_engine.cancellable_engine import CancellableEngine
 from voicevox_engine.core.core_initializer import initialize_cores
 from voicevox_engine.engine_manifest import load_manifest
 from voicevox_engine.preset.Preset import PresetManager
-from voicevox_engine.setting.Setting import (
-    USER_SETTING_PATH,
-    CorsPolicyMode,
-    SettingHandler,
-)
+from voicevox_engine.setting.model import CorsPolicyMode
+from voicevox_engine.setting.setting_manager import USER_SETTING_PATH, SettingHandler
 from voicevox_engine.tts_pipeline.tts_engine import make_tts_engines_from_cores
 from voicevox_engine.user_dict.user_dict import UserDictionary
 from voicevox_engine.utility.path_utility import engine_manifest_path, engine_root
